Refactored in preparation for adding new keyboards
