+++ common/appswitcher.c
