+++ common/combo_defs.h
