+++ common/custom_capsword.c
