+++ common/custom_edit.c
