+++ common/os_shortcut_defs.h
