+++ common/os_shortcuts.c
@@ -49,7 +49,7 @@ uint8_t os_get_raw() {
 
 const uint16_t PROGMEM os_shortcut_codes[][3] = {
     DEFINE_SHORTCUT( ID_START, KC_NO, KC_NO, KC_NO )
-#   include "os_shortcut_defs.h"
+#   include "../common/os_shortcut_defs.h"
     DEFINE_SHORTCUT( ID_END, KC_NO, KC_NO, KC_NO )
 };
 #endif
