+++ common/os_shortcuts.h
