+++ common/os_shortcuts_static/README.md
