+++ common/os_shortcuts_static/os_shortcuts_update.sh
