+++ common/shift_defs.h
