+++ kyria-rsthd-prime/rules.mk
@@ -18,6 +18,7 @@ GRAVE_ESC_ENABLE = no
 MAGIC_ENABLE = no
 MUSIC_ENABLE = no
 
+
 # frogm0uth features
 OS_SHORTCUTS = yes          # Enable OS shortcut mapping. If OS_SHORTCUTS_STATIC in config.h is
                             # defined, shortcuts will be statically compiled in. The script
@@ -35,27 +36,20 @@ CUSTOM_CAPSWORD = yes       # Enable custom capsword
 
 # Uncomment one of the next lines to prevent unused variable/function errors
 # from halting the compile
+#
+## NLR??
+#
 //EXTRAFLAGS += -Wno-unused-function -Wno-unused-variable
-EXTRAFLAGS += -Wno-error=unused-variable -Wno-error=unused-function
+//EXTRAFLAGS += -Wno-error=unused-variable -Wno-error=unused-function
 
 
 # Don't edit from here down
 
-SRC += appswitcher.c
-
-ifeq ($(strip $(CUSTOM_CAPSWORD)), yes)
-        SRC += custom_capsword.c
-        OPT_DEFS += -DCUSTOM_CAPSWORD
-endif
+SRC += ../common/process.c ../common/appswitcher.c ../common/util.c
 
 ifeq ($(strip $(OLED_ENABLE)), yes)
         SRC += oled.c
         OPT_DEFS += -DOLED_ENABLE
-endif
-
-ifeq ($(strip $(ENCODER_ENABLE)), yes)
-        SRC += encoder.c
-        OPT_DEFS += -DENCODER_ENABLE
 endif
 
 ifeq ($(strip $(RGBLIGHT_ENABLE)), yes)
@@ -63,28 +57,38 @@ ifeq ($(strip $(RGBLIGHT_ENABLE)), yes)
         OPT_DEFS += -DRGBLIGHT_ENABLE
 endif
 
+ifeq ($(strip $(ENCODER_ENABLE)), yes)
+        SRC += encoder.c
+        OPT_DEFS += -DENCODER_ENABLE
+endif
+
+ifeq ($(strip $(CUSTOM_CAPSWORD)), yes)
+        SRC += ../common/custom_capsword.c
+        OPT_DEFS += -DCUSTOM_CAPSWORD
+endif
+
 ifeq ($(strip $(COMBO_ENABLE)), yes)
-        SRC += combos.c
+        SRC += ../common/combos.c
         OPT_DEFS += -DCOMBO_ENABLE -DCOMBO_MUST_PRESS_IN_ORDER_PER_COMBO -DCOMBO_TERM_PER_COMBO
 endif
 
 ifeq ($(strip $(COMBOROLL_ENABLE)), yes)
-        SRC += comboroll.c
+        SRC += ../common/comboroll.c
         OPT_DEFS += -DCOMBOROLL_ENABLE
 endif
 
 ifeq ($(strip $(OS_SHORTCUTS)), yes)
-        SRC += os_shortcuts.c
+        SRC += ../common/os_shortcuts.c
         OPT_DEFS += -DOS_SHORTCUTS
 endif
 
 ifeq ($(strip $(CUSTOM_EDIT)), yes)
-        SRC += custom_edit.c
-        SRC += os_shortcuts.c
+        SRC += ../common/custom_edit.c
+        SRC += ../common/os_shortcuts.c
         OPT_DEFS += -DCUSTOM_EDIT -DOS_SHORTCUTS
 endif
 
 ifeq ($(strip $(LAYER_TAP_TOGGLE)), yes)
-        SRC += layer_tap_toggle.c
+        SRC += ../common/layer_tap_toggle.c
         OPT_DEFS += -DLAYER_TAP_TOGGLE
 endif
